feat: release package as a library
