+++ bin/index.ts
@@ -1,20 +1,10 @@
 #! /usr/bin/env node
-import { ApiClient, BugSplatApiClient, OAuthClientCredentialsClient, SymbolsApiClient, VersionsApiClient } from '@bugsplat/js-api-client';
+import { ApiClient, BugSplatApiClient, OAuthClientCredentialsClient, VersionsApiClient } from '@bugsplat/js-api-client';
 import commandLineArgs, { CommandLineOptions } from 'command-line-args';
 import commandLineUsage from 'command-line-usage';
-import { glob } from 'glob';
-import { existsSync } from 'node:fs';
-import { mkdir, readFile, stat } from 'node:fs/promises';
-import { basename, dirname, extname, join, relative } from 'node:path';
-import { pool } from 'workerpool';
+import { readFile, stat } from 'node:fs/promises';
+import { uploadSymbolFiles } from '../src/upload';
 import { CommandLineDefinition, argDefinitions, usageDefinitions } from './command-line-definitions';
-import { SymbolFileInfo } from './info';
-import { tryGetPdbGuid, tryGetPeGuid } from './pdb';
-import { getSymFileInfo } from './sym';
-import { safeRemoveTmp, tmpDir } from './tmp';
-import { createWorkersFromSymbolFiles } from './worker';
-
-const workerPool = pool(join(__dirname, 'compression.js'));
 
 (async () => {
     let {
@@ -75,11 +65,10 @@ const workerPool = pool(join(__dirname, 'compression.js'));
 
     console.log('Authentication success!');
 
-    const versionsApiClient = new VersionsApiClient(bugsplat);
-    const symbolsApiClient = new SymbolsApiClient(bugsplat);
-
     if (remove) {
         try {
+            const versionsApiClient = new VersionsApiClient(bugsplat);
+
             console.log(`About to delete symbols for ${database}-${application}-${version}...`);
 
             await versionsApiClient.deleteSymbols(
@@ -98,74 +87,14 @@ const workerPool = pool(join(__dirname, 'compression.js'));
     }
 
     directory = normalizeDirectory(directory);
-    const globPattern = `${directory}/${files}`;
 
-    let returnCode = 0;
-    try {
-        const symbolFilePaths = await glob(globPattern);
+    await uploadSymbolFiles(bugsplat, database, application, version, directory, files);
 
-        if (!symbolFilePaths.length) {
-            throw new Error(`Could not find any files to upload using glob ${globPattern}!`);
-        }
-
-        console.log(`Found files:\n ${symbolFilePaths.join('\n')}`);
-        console.log(`About to upload symbols for ${database}-${application}-${version}...`);
-
-        if (!existsSync(tmpDir)) {
-            await mkdir(tmpDir);
-        }
-
-        const symbolFiles = await Promise.all(symbolFilePaths.map(async (symbolFilePath) => await createSymbolFileInfo(directory, symbolFilePath)));
-        const workers = createWorkersFromSymbolFiles(workerPool, symbolFiles, [symbolsApiClient, versionsApiClient]);
-        const uploads = workers.map((worker) => worker.upload(database, application, version));
-        await Promise.all(uploads);
-
-        console.log('Symbols uploaded successfully!');
-    } catch (error) {
-        console.error(error);
-        returnCode = 1;
-    } finally {
-        await safeRemoveTmp();
-    }
-
-    process.exit(returnCode);
+    process.exit(0);
 })().catch((error) => {
     console.error(error.message);
     process.exit(1);
 });
-
-async function createSymbolFileInfo(searchDirectory: string, symbolFilePath: string): Promise<SymbolFileInfo> {
-    const path = symbolFilePath;
-    const relativePath = relative(searchDirectory, dirname(path));
-    const extLowerCase = extname(path).toLowerCase();
-    const isSymFile = extLowerCase.includes('.sym');
-    const isPdbFile = extLowerCase.includes('.pdb');
-    const isPeFile = extLowerCase.includes('.exe') || extLowerCase.includes('.dll');
-
-    let dbgId = '';
-    let moduleName = '';
-
-    if (isPdbFile) {
-        dbgId = await tryGetPdbGuid(path);
-    }
-
-    if (isPeFile) {
-        dbgId = await tryGetPeGuid(path);
-    }
-
-    if (isSymFile) {
-        ({ dbgId, moduleName } = await getSymFileInfo(path));
-    }
-
-    moduleName = moduleName || basename(path);
-
-    return {
-        path,
-        dbgId,
-        moduleName,
-        relativePath
-    } as SymbolFileInfo;
-}
 
 async function createBugSplatClient({
     user,
