+++ src/sym.ts
